rift: Add rift_evaluate_pose() and use it.
rift_evaluate_pose() is a unified place to project the
LEDs in a model and score them against a set of blobs.

This allows experimenting in a single place instead of spread
through the code.

The results of rift_evaluate_pose() are already better than
what preceded, because it counts blobs that map onto LEDs we
expect to be visible, instead of seeing if a projected LED
falls within any blob. The difference is that each blob gets
uniquely mapped to an LED, instead of allowing all LEDs to
potentially map onto a single blob.

Partially fixes https://github.com/thaytan/OpenHMD/issues/10