rift: Add device_state struct to frame processing.
Move the capture-time world pose for each device into a device_state
struct along with a final pose and score, so each frame can evaluate
each device independently.

This is a step toward doing the expensive correspondence search for
lost devices in a third thread, and also allows for the debug view to
show the final pose of the device when correspondence search wasn't
required - as previously it only retrieved the pose from the
correspondence search object.