rift: Quieten noisy "Lost delay slot" message.
There are only 3 fusion delay slots for each device - one for the
exposure we're expecting, one for the exposure we're capturing / doing
fast analysis on and one for a frame that's undergoing long analysis.

If either analysis phase takes too long, we can run out of free fusion
slots temporarily.