rift: Change the HMD IMU calibration to match controllers
Apply the rotation matrix, then the calibration offsets as
for controllers. The difference is marginal, because the
rotation matrix is near identity, but this seems more
correct based on monitoring raw IMU outputs.

Also fixes a bug where the rotation matrix calculation
was modifying the input x/y/z as it was calculating the
outputs.