rift: Change the way the tracker accepts orientation observations.
A common failure seems to be that we've lost track of the yaw, and
therefore orientatino fails to match. To avoid that, shorten
the threshold for forced orientation updates. On balance this
seems to reject short-term spurious poses well, while still
accepting long term trends, but I think it could be better still.