rift: Don't invalidate a tracking slot on frame release.
Now that exposures are associated only at frame capture time,
instead of at the start of frame arrivals, it's likely that
we can capture a frame and completely analyse it before any
other sensor completes capture. In that case, we don't want
to claim the delay slot and then release it (making it
unavailable for later frames). Instead, just reduce the
use_count to 0 so that the delay slot may be reused later,
but is still available if other frames do arrive in time.